+++ src/lib/types/CommonInputProps.ts
@@ -1,7 +1,13 @@
 import { ReactNode, CSSProperties } from "react";
 import { FieldValues } from "react-hook-form";
 
-interface CommonInputProps<T extends FieldValues> {
+interface DefaultAddonProps {
+  isDisabled?: boolean;
+}
+
+export type MergedAddonProps<TRenderAddon> = TRenderAddon & DefaultAddonProps;
+
+interface CommonInputProps<T extends FieldValues, TRenderAddon = unknown> {
   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
   onBlur?: (e: React.FocusEvent<HTMLInputElement>) => void;
   label?: ReactNode;
@@ -33,6 +39,7 @@ interface CommonInputProps<T extends FieldValues> {
   /**
    * Component prop that represents an element rendered inside the input group on the left side: https://reactstrap.github.io/?path=/docs/components-inputgroup--input-group
    * If you want to render text or an icon inside the input, it is recommended to wrap it inside an `InputGroupText` component.
+   * The `addonLeft` prop also accepts a function that receives various props based on the input type and returns a ReactNode.
    * @type {React.ReactNode}
    * @example <caption>Render a string by wrapping it inside a `InputGroupText` component</caption>
    * addonLeft={<InputGroupText>@</InputGroupText>}
@@ -40,11 +47,12 @@ interface CommonInputProps<T extends FieldValues> {
    * @example <caption>Render a `Button` component</caption>
    * addonLeft={<Button>Button</Button>}
    */
-  addonLeft?: React.ReactNode;
+  addonLeft?: React.ReactNode | ((props: MergedAddonProps<TRenderAddon>) => React.ReactNode);
 
   /**
    * Component prop that represents an element rendered inside the input group on the right side: https://reactstrap.github.io/?path=/docs/components-inputgroup--input-group
    * If you want to render text or an icon inside the input, it is recommended to wrap it inside an `InputGroupText` component.
+   * The `addonRight` prop also accepts a function that receives various props based on the input type and returns a ReactNode.
    * @type {React.ReactNode}
    * @example <caption>Render a string by wrapping it inside a `InputGroupText` component</caption>
    * addonRight={<InputGroupText>@</InputGroupText>}
@@ -52,7 +60,7 @@ interface CommonInputProps<T extends FieldValues> {
    * @example <caption>Render a `Button` component</caption>
    * addonRight={<Button>Button</Button>}
    */
-  addonRight?: React.ReactNode;
+  addonRight?: React.ReactNode | ((props: MergedAddonProps<TRenderAddon>) => React.ReactNode);
 
   /**
    * Component prop that represents an additional style attribute for the input group